+++ .editorconfig
@@ -1,0 +1,8 @@
+root = true
+
+[{*.kt,*.kts}]
+ij_kotlin_allow_trailing_comma_on_call_site=true
+ij_kotlin_allow_trailing_comma=true
+max_line_length=100
+ij_kotlin_packages_to_use_import_on_demand = java.util.*,kotlinx.android.synthetic.**,org.junit.Assert.*
+
