+++ build.gradle.kts
@@ -17,33 +17,33 @@ subprojects {
     apply(plugin = "org.jetbrains.dokka")
 
     apply(plugin = "org.jlleitschuh.gradle.ktlint")
-    ktlint {
-        android.set(true)
-        verbose.set(true)
-        debug.set(true)
-        outputToConsole.set(true)
-        outputColorName.set("RED")
-        ignoreFailures.set(false)
-        enableExperimentalRules.set(true)
-        disabledRules.set(
-            setOf(
-                "no-wildcard-imports",
-                "filename",
-                "package-name",
-                "experimental:package-name",
-                "experimental:comment-wrapping",
-            ),
-        )
-        reporters {
-            reporter(org.jlleitschuh.gradle.ktlint.reporter.ReporterType.PLAIN)
-            reporter(org.jlleitschuh.gradle.ktlint.reporter.ReporterType.CHECKSTYLE)
-            reporter(org.jlleitschuh.gradle.ktlint.reporter.ReporterType.SARIF)
-        }
-        kotlinScriptAdditionalPaths {
-            include(fileTree("scripts/"))
-        }
-        filter {
-            exclude { element -> element.file.path.contains("generated/") }
-        }
-    }
+//    ktlint {
+//        android.set(true)
+//        verbose.set(true)
+//        debug.set(true)
+//        outputToConsole.set(true)
+//        outputColorName.set("RED")
+//        ignoreFailures.set(false)
+//        enableExperimentalRules.set(true)
+//        disabledRules.set(
+//            setOf(
+//                "no-wildcard-imports",
+//                "filename",
+//                "package-name",
+//                "experimental:package-name",
+//                "experimental:comment-wrapping",
+//            ),
+//        )
+//        reporters {
+//            reporter(org.jlleitschuh.gradle.ktlint.reporter.ReporterType.PLAIN)
+//            reporter(org.jlleitschuh.gradle.ktlint.reporter.ReporterType.CHECKSTYLE)
+//            reporter(org.jlleitschuh.gradle.ktlint.reporter.ReporterType.SARIF)
+//        }
+//        kotlinScriptAdditionalPaths {
+//            include(fileTree("scripts/"))
+//        }
+//        filter {
+//            exclude { element -> element.file.path.contains("generated/") }
+//        }
+//    }
 }
