+++ compose-ui/src/main/java/com/flexcode/multiselectcalendar/composeUi/theme/Theme.kt
@@ -1,4 +1,4 @@
-package com.flexcode.multiselectcalendar.compose_ui.theme
+package com.flexcode.multiselectcalendar.composeUi.theme
 
 import android.app.Activity
 import android.os.Build
