+++ compose-ui/src/main/java/com/flexcode/multiselectcalendar/composeUi/theme/Type.kt
@@ -1,4 +1,4 @@
-package com.flexcode.multiselectcalendar.compose_ui.theme
+package com.flexcode.multiselectcalendar.composeUi.theme
 
 import androidx.compose.material3.Typography
 import androidx.compose.ui.text.TextStyle
@@ -6,7 +6,7 @@ import androidx.compose.ui.text.font.Font
 import androidx.compose.ui.text.font.FontFamily
 import androidx.compose.ui.text.font.FontWeight
 import androidx.compose.ui.unit.sp
-import com.flexcode.multiselectcalendar.compose_ui.R
+import com.flexcode.multiselectcalendar.composeUi.R
 
 // Set of Material typography styles to start with
 val Typography = Typography(
