+++ lib/src/main/java/com/flexcode/multiselectcalendar/utils/ExperimentalMultiSelectCalendarApi.kt
@@ -2,7 +2,7 @@ package com.flexcode.multiselectcalendar.utils
 
 @RequiresOptIn(
     level = RequiresOptIn.Level.WARNING,
-    message = "This is an experimental API. It may be changed or removed in the future.",
+    message = "This is an experimental API. It may be unstable",
 )
 @Retention(AnnotationRetention.BINARY)
 @Target(AnnotationTarget.CLASS, AnnotationTarget.FUNCTION)
