+++ lib/src/main/java/com/flexcode/multiselectcalendar/utils/WeekUtil.kt
@@ -50,3 +50,16 @@ internal fun YearMonth.getWeeks(firstDayOfTheWeek: DayOfWeek, today: LocalDate):
         )
     }
 }
+
+internal fun Collection<LocalDate>.addOrRemoveIfExists(date: LocalDate) = if (contains(date)) {
+    this - date
+} else {
+    this + date
+}
+
+internal infix fun DayOfWeek.daysUntil(other: DayOfWeek) =
+    if (Build.VERSION.SDK_INT >= Build.VERSION_CODES.O) {
+        (7 + (value - other.value)) % 7
+    } else {
+        throw CustomException()
+    }
